+++ pytorch_lightning/strategies/single_tpu.py
@@ -47,8 +47,6 @@ class SingleTPUStrategy(SingleDeviceStrategy):
         )
 
         self.debug = debug
-        self.tpu_local_core_rank = 0
-        self.tpu_global_core_rank = 0
 
     @property
     def is_distributed(self) -> bool:
@@ -66,9 +64,6 @@ class SingleTPUStrategy(SingleDeviceStrategy):
 
         if self.debug:
             os.environ["PT_XLA_DEBUG"] = str(1)
-
-        self.tpu_local_core_rank = xm.get_local_ordinal()
-        self.tpu_global_core_rank = xm.get_ordinal()
 
     def model_to_device(self) -> None:
         self.model.to(self.root_device)